[953] Refactor: move time since calculations to ind_cqc_pipeline
