+++ projects/_01_ingest/cqc_api/jobs/delta_clean_cqc_location_data.py
@@ -18,7 +18,6 @@ from utils.column_names.cleaned_data_files.cqc_location_cleaned import (
 )
 from utils.column_values.categorical_column_values import (
     CareHome,
-    Dormancy,
     LocationType,
     PrimaryServiceType,
     RegistrationStatus,
@@ -108,8 +107,6 @@ def main(
     cqc_location_df = cUtils.column_to_date(
         cqc_location_df, Keys.import_date, CQCLClean.cqc_location_import_date
     )
-    # cqc_location_df = calculate_time_registered_for(cqc_location_df)
-    # cqc_location_df = calculate_time_since_dormant(cqc_location_df)
 
     cqc_location_df = impute_historic_relationships(cqc_location_df)
     registered_locations_df = select_registered_locations_only(cqc_location_df)
@@ -643,64 +640,6 @@ def select_registered_locations_only(locations_df: DataFrame) -> DataFrame:
         locations_df[CQCL.registration_status] == RegistrationStatus.registered
     )
     return locations_df
-
-
-def calculate_time_since_dormant(df: DataFrame) -> DataFrame:
-    """
-    Adds a column to show the number of months since the location was last dormant.
-
-    This function calculates the number of months since the last time a location was marked as dormant.
-    It uses a window function to track the most recent date when dormancy was marked as "Y" and calculates
-    the number of months since that date for each location.
-
-    'time_since_dormant' values before the first instance of dormancy are null.
-    If the location has never been dormant then 'time_since_dormant' is null.
-
-    Args:
-        df (DataFrame): A dataframe with columns: cqc_location_import_date, dormancy, and location_id.
-
-    Returns:
-        DataFrame: A dataframe with an additional column 'time_since_dormant'.
-    """
-    w = Window.partitionBy(CQCLClean.location_id).orderBy(
-        CQCLClean.cqc_location_import_date
-    )
-
-    df = df.withColumn(
-        CQCLClean.dormant_date,
-        F.when(
-            F.col(CQCLClean.dormancy) == Dormancy.dormant,
-            F.col(CQCLClean.cqc_location_import_date),
-        ),
-    )
-
-    df = df.withColumn(
-        CQCLClean.last_dormant_date,
-        F.last(CQCLClean.dormant_date, ignorenulls=True).over(w),
-    )
-
-    df = df.withColumn(
-        CQCLClean.time_since_dormant,
-        F.when(
-            F.col(CQCLClean.last_dormant_date).isNotNull(),
-            F.when(F.col(CQCLClean.dormancy) == Dormancy.dormant, 1).otherwise(
-                F.floor(
-                    F.months_between(
-                        F.col(CQCLClean.cqc_location_import_date),
-                        F.col(CQCLClean.last_dormant_date),
-                    )
-                )
-                + 1,
-            ),
-        ),
-    )
-
-    df = df.drop(
-        CQCLClean.dormant_date,
-        CQCLClean.last_dormant_date,
-    )
-
-    return df
 
 
 def add_cqc_sector_column_to_cqc_locations_dataframe(
