+++ utils/column_names/ind_cqc_pipeline_columns.py
@@ -338,9 +338,9 @@ class IndCqcColumns:
         "sum_non_rm_managerial_estimated_filled_posts"
     )
     time_between_submissions: str = "time_between_submissions"
-    time_registered: str = CQCLClean.time_registered
+    time_registered: str = "time_registered"
     time_registered_capped_at_four_years: str = "time_registered_capped_at_four_years"
-    time_since_dormant: str = CQCLClean.time_since_dormant
+    time_since_dormant: str = "time_since_dormant"
     total_staff_bounded: str = AWPClean.total_staff_bounded
     unix_time: str = "unix_time"
     upper_percentile: str = "upper_percentile"
